+++ R/auk-filter.r
@@ -58,7 +58,7 @@
 #' filters <- auk_ebd(f) %>%
 #'   auk_species(species = c("Gray Jay", "Blue Jay")) %>%
 #'   auk_country(country = c("US", "Canada")) %>%
-#'   auk_extent(extent = c(-100, 37, -80, 52)) %>%
+#'   auk_bbox(bbox = c(-100, 37, -80, 52)) %>%
 #'   auk_date(date = c("2012-01-01", "2012-12-31")) %>%
 #'   auk_time(start_time = c("06:00", "09:00")) %>%
 #'   auk_duration(duration = c(0, 60)) %>%
@@ -68,7 +68,7 @@
 #' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
 #' filters <- auk_species(ebd, species = c("Gray Jay", "Blue Jay"))
 #' filters <- auk_country(filters, country = c("US", "Canada"))
-#' filters <- auk_extent(filters, extent = c(-100, 37, -80, 52))
+#' filters <- auk_bbox(filters, bbox = c(-100, 37, -80, 52))
 #' filters <- auk_date(filters, date = c("2012-01-01", "2012-12-31"))
 #' filters <- auk_time(filters, start_time = c("06:00", "09:00"))
 #' filters <- auk_duration(filters, duration = c(0, 60))
@@ -415,9 +415,9 @@ awk_translate <- function(filters, col_idx, sep, select) {
                         collapse = " || ")
     filter_strings$bcr <- str_interp(awk_if, list(condition = condition))
   }
-  # extent filter
-  if (length(filters$extent) == 0) {
-    filter_strings$extent <- ""
+  # bbox filter
+  if (length(filters$bbox) == 0) {
+    filter_strings$bbox <- ""
   } else {
     lat_idx <- col_idx$index[col_idx$id == "lat"]
     lng_idx <- col_idx$index[col_idx$id == "lng"]
@@ -426,9 +426,9 @@ awk_translate <- function(filters, col_idx, sep, select) {
                         "$${lat_idx} >= ${ymn} && ",
                         "$${lat_idx} <= ${ymx}") %>%
       str_interp(list(lat_idx = lat_idx, lng_idx = lng_idx,
-                      xmn = filters$extent[1], xmx = filters$extent[3],
-                      ymn = filters$extent[2], ymx = filters$extent[4]))
-    filter_strings$extent <- str_interp(awk_if, list(condition = condition))
+                      xmn = filters$bbox[1], xmx = filters$bbox[3],
+                      ymn = filters$bbox[2], ymx = filters$bbox[4]))
+    filter_strings$bbox <- str_interp(awk_if, list(condition = condition))
   }
   # date filter
   if (length(filters$date) == 0) {
@@ -565,7 +565,7 @@ BEGIN {
   ${country}
   ${state}
   ${bcr}
-  ${extent}
+  ${bbox}
   ${date_substr}
   ${date}
   ${time}