+++ tests/testthat/test_filters.r
@@ -134,30 +134,30 @@ test_that("auk_bcr", {
   expect_error(auk_bcr(ebd, 100))
 })
 
-test_that("auk_extent", {
+test_that("auk_bbox", {
   ebd <- system.file("extdata/ebd-sample.txt", package = "auk") %>%
     auk_ebd()
 
   # works correctly
   e <- c(-125, 37, -120, 52)
-  ebd <- auk_extent(ebd, e)
-  expect_equal(ebd$filters$extent, e)
+  ebd <- auk_bbox(ebd, e)
+  expect_equal(ebd$filters$bbox, e)
 
   # overwrite
   e <- c(0, 0, 1, 1)
-  ebd <- auk_extent(ebd, e)
-  expect_equal(ebd$filters$extent, e)
+  ebd <- auk_bbox(ebd, e)
+  expect_equal(ebd$filters$bbox, e)
 
   # invalid lat
-  expect_error(auk_extent(ebd, c(0, -91, 1, 1)))
-  expect_error(auk_extent(ebd, c(0, -90, 1, 91)))
-  expect_error(auk_extent(ebd, c(0, 1, 1, 0)))
-  expect_error(auk_extent(ebd, c(0, 0, 1, 0)))
+  expect_error(auk_bbox(ebd, c(0, -91, 1, 1)))
+  expect_error(auk_bbox(ebd, c(0, -90, 1, 91)))
+  expect_error(auk_bbox(ebd, c(0, 1, 1, 0)))
+  expect_error(auk_bbox(ebd, c(0, 0, 1, 0)))
   # invalid lng
-  expect_error(auk_extent(ebd, c(-181, 0, 1, 1)))
-  expect_error(auk_extent(ebd, c(-180, 0, 181, 1)))
-  expect_error(auk_extent(ebd, c(1, 0, 0, 1)))
-  expect_error(auk_extent(ebd, c(0, 0, 0, 1)))
+  expect_error(auk_bbox(ebd, c(-181, 0, 1, 1)))
+  expect_error(auk_bbox(ebd, c(-180, 0, 181, 1)))
+  expect_error(auk_bbox(ebd, c(1, 0, 0, 1)))
+  expect_error(auk_bbox(ebd, c(0, 0, 0, 1)))
 })
 
 test_that("auk_date", {